+++ randovania/games/prime1/game_data.py
@@ -28,10 +28,10 @@ def _gui() -> game.GameGui:
 def _generator() -> game.GameGenerator:
     from randovania.games.prime1.generator.base_patches_factory import PrimeBasePatchesFactory
     from randovania.games.prime1.generator.bootstrap import PrimeBootstrap
-    from randovania.games.prime1.generator.item_pool.pool_creator import prime1_specific_pool
+    from randovania.games.prime1.generator.pickup_pool.pool_creator import prime1_specific_pool
 
     return game.GameGenerator(
-        item_pool_creator=prime1_specific_pool,
+        pickup_pool_creator=prime1_specific_pool,
         bootstrap=PrimeBootstrap(),
         base_patches_factory=PrimeBasePatchesFactory(),
         hint_distributor=PrimeHintDistributor(),