+++ randovania/games/super_metroid/game_data.py
@@ -27,7 +27,7 @@ def _generator() -> game.GameGenerator:
     from randovania.resolver.bootstrap import Bootstrap
 
     return game.GameGenerator(
-        item_pool_creator=super_metroid_specific_pool,
+        pickup_pool_creator=super_metroid_specific_pool,
         bootstrap=Bootstrap(),
         base_patches_factory=BasePatchesFactory(),
         hint_distributor=AllJokesHintDistributor(),