+++ test/conftest.py
@@ -17,7 +17,6 @@ from randovania.game_description.resources.pickup_entry import PickupEntry, Pick
 from randovania.game_description.resources.resource_database import ResourceDatabase
 from randovania.games import default_data
 from randovania.games.blank.layout import BlankConfiguration
-from randovania.games.cave_story.layout.cs_configuration import CSConfiguration
 from randovania.games.game import RandovaniaGame
 from randovania.games.prime1.layout.prime_configuration import PrimeConfiguration
 from randovania.games.prime2.exporter.game_exporter import decode_randomizer_data
@@ -130,17 +129,6 @@ def default_prime_configuration() -> PrimeConfiguration:
 @pytest.fixture()
 def prime_game_patches(default_prime_configuration, prime_game_description) -> GamePatches:
     return GamePatches.create_from_game(prime_game_description, 0, default_prime_configuration)
-
-
-@pytest.fixture(scope="session")
-def default_cs_preset() -> Preset:
-    return PresetManager(None).default_preset_for_game(RandovaniaGame.CAVE_STORY).get_preset()
-
-
-@pytest.fixture(scope="session")
-def default_cs_configuration(default_cs_preset) -> CSConfiguration:
-    assert isinstance(default_cs_preset.configuration, CSConfiguration)
-    return default_cs_preset.configuration
 
 
 @pytest.fixture(scope="session")